Refactor GUI pages into separate modules
